Refactoring of scenario and load tests
